remove standalone test, include test case in regressions #274
