--- tests/test_readonly_related_field.yml
@@ -1,222 +1,0 @@
-openapi: 3.0.3
-info:
-  title: ''
-  version: 0.0.0
-paths:
-  /referencing_model/:
-    get:
-      operationId: referencing_model_list
-      description: ''
-      tags:
-      - referencing_model
-      security:
-      - cookieAuth: []
-      - basicAuth: []
-      - {}
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/ReferencingModel'
-          description: ''
-    post:
-      operationId: referencing_model_create
-      description: ''
-      tags:
-      - referencing_model
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/ReferencingModel'
-          application/x-www-form-urlencoded:
-            schema:
-              $ref: '#/components/schemas/ReferencingModel'
-          multipart/form-data:
-            schema:
-              $ref: '#/components/schemas/ReferencingModel'
-        required: true
-      security:
-      - cookieAuth: []
-      - basicAuth: []
-      responses:
-        '201':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ReferencingModel'
-          description: ''
-  /referencing_model/{id}/:
-    get:
-      operationId: referencing_model_retrieve
-      description: ''
-      parameters:
-      - in: path
-        name: id
-        schema:
-          type: string
-          format: uuid
-        description: A UUID string identifying this referencing model.
-        required: true
-      tags:
-      - referencing_model
-      security:
-      - cookieAuth: []
-      - basicAuth: []
-      - {}
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ReferencingModel'
-          description: ''
-    put:
-      operationId: referencing_model_update
-      description: ''
-      parameters:
-      - in: path
-        name: id
-        schema:
-          type: string
-          format: uuid
-        description: A UUID string identifying this referencing model.
-        required: true
-      tags:
-      - referencing_model
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/ReferencingModel'
-          application/x-www-form-urlencoded:
-            schema:
-              $ref: '#/components/schemas/ReferencingModel'
-          multipart/form-data:
-            schema:
-              $ref: '#/components/schemas/ReferencingModel'
-        required: true
-      security:
-      - cookieAuth: []
-      - basicAuth: []
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ReferencingModel'
-          description: ''
-    patch:
-      operationId: referencing_model_partial_update
-      description: ''
-      parameters:
-      - in: path
-        name: id
-        schema:
-          type: string
-          format: uuid
-        description: A UUID string identifying this referencing model.
-        required: true
-      tags:
-      - referencing_model
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/PatchedReferencingModel'
-          application/x-www-form-urlencoded:
-            schema:
-              $ref: '#/components/schemas/PatchedReferencingModel'
-          multipart/form-data:
-            schema:
-              $ref: '#/components/schemas/PatchedReferencingModel'
-      security:
-      - cookieAuth: []
-      - basicAuth: []
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ReferencingModel'
-          description: ''
-    delete:
-      operationId: referencing_model_destroy
-      description: ''
-      parameters:
-      - in: path
-        name: id
-        schema:
-          type: string
-          format: uuid
-        description: A UUID string identifying this referencing model.
-        required: true
-      tags:
-      - referencing_model
-      security:
-      - cookieAuth: []
-      - basicAuth: []
-      responses:
-        '204':
-          description: No response body
-components:
-  schemas:
-    PatchedReferencingModel:
-      type: object
-      properties:
-        id:
-          type: string
-          format: uuid
-          readOnly: true
-        referenced_model:
-          type: integer
-        referenced_model_readonly:
-          type: integer
-          readOnly: true
-        referenced_model_m2m:
-          type: array
-          items:
-            type: integer
-        referenced_model_m2m_readonly:
-          type: array
-          items:
-            type: integer
-          readOnly: true
-    ReferencingModel:
-      type: object
-      properties:
-        id:
-          type: string
-          format: uuid
-          readOnly: true
-        referenced_model:
-          type: integer
-        referenced_model_readonly:
-          type: integer
-          readOnly: true
-        referenced_model_m2m:
-          type: array
-          items:
-            type: integer
-        referenced_model_m2m_readonly:
-          type: array
-          items:
-            type: integer
-          readOnly: true
-      required:
-      - id
-      - referenced_model
-      - referenced_model_m2m
-      - referenced_model_m2m_readonly
-      - referenced_model_readonly
-  securitySchemes:
-    basicAuth:
-      type: http
-      scheme: basic
-    cookieAuth:
-      type: apiKey
-      in: cookie
-      name: Session
